Spread out a given node's control messages over real time:
Now that Sonic Pi operates using virtual/logical timestamps which mechanically
represent real-time,  when between two adjacent sleeps, time is assumed to be static. 
This has the effect that all control messages in such a region share the same time stamp
which means that the actual handling of the control messages at the server is
non-deterministic. We now add a 5ms delay between subsequent control messages for a
specific node on a specific thread. This preserves the logical ordering of the 
control messages at a slight cost of incremental (in 5ms chunks) latency.

One drawback of this approach is that if a synth node is shared between
threads, these logical deltas will no longer be honoured, and non-determinism 
will return...