Fix mismatched input/output sample rates on OSX
Fixes issue #87
SuperCollider can't boot if the sample rates do
not match. It can however boot at the second time
of asking if a sample rate is specified. The check
for sample rates only works on OSX 10.7 and later
as system_profiler didn't include sample rate info
until that version. In the case of 10.6 users, the
default behaviour will now be to log a warning and
fail to boot if SuperCollider can't start.