Clarify Qt install process for OSX
When following the instructions, several users found that the Qt app
hung at the splash screen after compiling it themselves. This is due to
their installs not having a ruby binary in the right location, which is
hardcoded into the Qt app. This PR suggests a workaround in the way of
symlinking an appropriate Ruby into the correct location, and generally
tries to clear up the steps in INSTALL.md

The changes regarding `install_name_tool` are to prevent the following
error:

    ERROR: no file at "/usr/lib/libqscintilla2.11.dylib"

whilst running the `mac-build-app` script. This is because `macdeployqt`
gets confused unless the inner path of the qscintilla dylib is a full
path. More information about this error is written up in issue #123.
This PR fixes #123.