context-sensitive F1
F1 opens a help page depending on the selection, or the word the cursor is
in/after.  Help page titles are used as the keywords, but
pre-titleization, so ":dull_bell" works to look up the "Dull Bell" help
page.  Rerun qt-doc.rb before building since struct help_page has changed
to add the keyword field.