Added instruction for running under generic linux
JACK doesn't start automatically when running the script. Added instructions for starting it manually. 

Configuring JACK for low-latency performance is not difficult, and there are several guides for how to do this online, but I didn't know a good up to date walk-through to point to. The jackaudio [wiki](https://github.com/jackaudio/jackaudio.github.com/wiki) might be a good choice?