Try interpreting missing methods as symbols
Rather than erroring out for a missing method, first try interpreting it as a symbol.
This may be the desired behavior for many cases, especially with those new to Ruby.
For example, the following will do the expected thing:

  with_synth dsaw do
    play_pattern [c, d]
    sample loop_amen
  end

and it also reads more like a proper DSL.
This usage shouldn't be encouraged, since method_missing has a nontrivial overhead,
but it is better than erroring out when you're trying to make music!