Teach live loop to check for sleeps:
* If no sleeps happened within the live loop, raise
  an exception. 

This fixes an annoying issue I was having during a
gig when I was defining a function with a chord
and then in a live loop I iterated through each note, 
playing then sleeping. However, when I accidentally
defined the function to return nil, the iteration 
never took place and the thread just started thrashing 
round in a fast loop stealing resources and forcing 
a restart.