Give each run a new rand generator solely for new threads:
Previously the standard thread local random generator
was used to create a new generator for each thread
in a deterministic manner. However, if multiple threads
were created sequentially the previous approach gave
each thread a similar generator (i.e. with the same seed).

This new approach uses a separate generator for generating
new thread generator seeds. This allows the standard 
generator to not be polluted with new generator generations
in addition to giving each new thread's generator
a new seed in a deterministic manner.