History: Add links to Github tags for each version
Each version in History now has a link to all the
individual commits that were part of its development
on Github. This should make the path to exploring
the ongoing development of Sonic Pi a little easier
to find.

I'm linking to the commit history rather than the
specific tag page to jumpstart the discovery that
the history of open source software development is 
completely available for anyone (provided you use 
a version control system such as git).