OS X - Force sample rate to 44k
If the sample rate of the default input source
doesn't match the sample rate of the default output
source, then scsynth refuses to boot.

On boot, we now force the sample rate of both default 
input and output to 44100 to ensure they're identical.

This is enabled through some native extensions in
the library ruby-coreaudio.

This only applies for installations on OS X.

Other modifications:

* both bin/compile-extensions and bin/sonic-pi-server
  have now gained a new top level var os which 
  contains a symbol of the current OS. This 
  functionality is duplicated, so consider a way 
  of unifying this.

* all contents of the rb-native subdirectory specific
  to the OS and Ruby version are now added to
  the library path so all compiled native libraries
  are available for requiring.

Fixes: #273, #87