Flanger Info - tidy up param docstrings
The ranges of the params have changed and some are
no longer 0 - 1.

I really wish I knew how to better describe these
params to communicate their meaning. One one hand
it's nice to leave things open for experimentation.
On another it would be nice to guide the initial
experience.

* Perhaps we really need an examples section for
both FX and Synths.

* Perhaps also a param library of maps which can
  act as named presets.