Add deprecation warning for with_tempo
In v1.0 with_tempo was used to set the current
bpm. However, v2.0 unified around the style of
use_* for setting the current state in the current
thread and with_* for doing the same but only for
the context of the block. Also, the term tempo
was replaced with bpm.

There is currently a number of known pieces of 
printed literature which still reference the v1.0
syntax - so provide users working with that material
a nice way to know that the syntax has evolved.