Breaking Change in API - range is now exclusive
range no longer returns a ring inclusive of the
finish value.

Previously:

(range 1, 5) #=> (ring 1, 2, 3, 4, 5)

Now:

(range 1, 5) #=> (ring 1, 2, 3, 4)

This is to correctly map onto the Clojure semantics 
which this fn was originally modelled on. Also,
not including the finish value allows range to 
more effectively be used for things that wrap
such as time:

at (range 0, 1, 0.125) do
  # do something
end

Previously, the block would be executed twice each
time the beat wrapped round - at 0 and at 1.