Breaking change - thread rand gen is now reset alongside standard gen
When new threads are created with in_thread, they
get a brand new random generator which all rand
calls (i.e. choose, rand, rrand and friends) use
to be random. Each thread has a separate rand generator
to isolate randomisation within each thread and to stop
threads affecting the randomisation of other threads
removing races and therefore making the randomisation 
behaviour deterministic across runs.

Sometimes it is interesting to reset the random
generator to deterministically create patterns from
random streams:

loop do
  use_random_seed 700
  notes = (scale :e3, :minor)
  16.times do
    play notes.choose, release: 0.1
    sleep 0.125
  end
end

The above code will play the same melody of 16
notes repeatedly.

Previously use_random_seed and its friend with_random_seed
only reset the standard random generator. However,
the random generator used for new threads was left
unaffected. This meant the following code produced
a random pattern of sample triggers as the call to
use_random_seed did not affect the rand gen used to
generate new generators for new threads:

live_loop :kik do
  use_random_seed 700
  at (range 0, 1, 0.125 / 1) do
    r = [1, 2, 3, 4, 5].choose
    puts r
    sample :sn_dolf, rate: 8 if r < 5
  end
  sleep 1
end

Now, when the standard rand generator is reset, 
the generator used to create new rand generators
is reset too as a function of the current rand gen.
This means that the reset is deterministic and also
retains the property that the rand gen of a new thread
is not the same as the rand gen of its parent thread.