Teach FX to combine args before determining kill_delay
FX are allowed to override the fn kill_delay to
programmatically determine the amount of time to wait
after all the contents of their block has completed
before killing themselves. The default is just 1s, 
however, some FX may wish to make this time a function
of some of their arguments. For example, the reverb
FX benefits from a longer kill_delay for larger room
sizes, otherwise the FX is killed before the sound
has completed.

A previous commit created this relationship. However,
the room param was not available at the time the
kill_delay fn was called as it was created as part
of the process of triggering synths.

This synth param combination and resolution 
functionality has been re-organised to enable the
with_fx to have access to the final synth args
so that the kill_delay fn (if defined) may work
correctly.

As a result, the fn normalise_and_resolve_synth_args
now takes an extra optional parameter combine_tls
which determines whether or not to merge in thread 
local synth params. This is now the case for calls
to play and synth. However, samples and FX do not
pull in and merge these thread locals.

with_fx now performs the normalisation and resolution
work before triggering the FX synth, and now uses
the bypass fn trigger_synth_with_resolved_args
to avoid normalising twice (and potentially calling any 
internal lambdas multiple times)

Fixes issue with reverb requiring an explicit room param.