Modify Tilburg example to run smoothly on a RP
The piece has been optimised by moving the reverb
FX creation outside of the live loop - this means 
only one reverb is created and used for the entire
live loop. The cost of this is that it's no longer
possible to modify the reverb as it's playing. 

Also modulate the pitch of the zawa synth with a
knitted pattern.