Remember last seen server version number
Given that we are now only checking for new versions
every 2 weeks, remember the last version seen
and use that to report whether or not the current
version is outdated when we're not actually asking
the server.

We're remembering the last seen version using the
new settings system.

If there is no last seen version, or if it's malformed,
we just return version 0.0.0