Add `spark` method for quick graphing of numbers
This adds a small method to render a UTF-8 based spark graph to a
string, suitable for `puts`ing to standard out. It takes a list of
numeric arguments or a single array of numeric arguments. For nested
arrays or non-numeric lists it returns a string saying `"spark
error:..."`

See the tests in `test_spark.rb` for expected behaviour.

Implementation stolen from this gist:
https://gist.github.com/jcromartie/1367091