Fix multibyte chars in OSC implementation
As of Ruby 1.9 the String#size method rationalised itself for multibyte
chars, so that a unicode string of a single character would have a size
of 1. This was a sensible fix *unless* you were doing something that had
to take account of how many bytes were used to represent the character.
This is what was wrong with our OSC implementation (and was actually an
issue against one of the ruby OSC libs - see maca/ruby-osc#4). This
meant that characters that took 3 bytes to represent didn't work nicely.

This was very confusing to debug as the spark function appeared to work
fine on OSX. Then it didn't on the Raspberry Pi (for which the Unicode
str function is added here). But spark didn't work for arrays of odd
lengths

```
spark(1,2) # worked
spark(1,2,3) # broke the log output
```

This was because the [utf8 block
chars](http://www.utf8-character.info/#!%E2%96%81%20%E2%96%82%20%E2%96%83%20%E2%96%84%20%E2%96%85%20%E2%96%86%20%E2%96%87)
used in spark take 3 bytes to represent. Therefore pairing them up to
form chains of 6's (even numbers of bytes) worked, whereas odd numbers
left 3 trailing bytes which broke the modulo function that relied on
String#size. In short, this was a hell of a bug.