Add synth_violin
Based on [Roger Allen's Overtone
patch](https://gist.githubusercontent.com/rogerallen/5992549/raw/2e4ed49bef990817e83981d812ab609e1b3bb901/violin.clj) which is, in turn, based on an
article in Sound on Sound magazine.

The resulting synth is a bit cheesy but sounds convincing enough in the
upper register with a little reverb. The main difference between this
and a raw saw wave is a set of "formant" filters (band pass filters at
particular frequencies) that mimic the resonances of a violin body
(supposedly...).

The other difference is the addition of a vibrato which can be
controlled through the params. As with a real violin, it takes practice
and good taste to get the vibrato to sound right!