WIP: Idea for adding probability to slicer fx
This is a proof of concept as we'd want any randomness in ugens to be
controlled from SonicPi/Ruby's random_seed generator to be reproducible
for each thread.

The design here works ok for everything except the sawtooth waveform
which is tricky to sample at exactly the point where it's silent. This
means that the sawtooth slicer just sounds like it's changing volume on
the beat instead of slicing.

This is all a development of the idea of adding a degrade function to
SP, as detailed in this gist

https://gist.github.com/xavriley/5d7859a4554e8f303b68