WIP: Add line marker for current line
This adds a line marker into the GUI for `play` and `sample` commands.
It probably needs it's own osc endpoint. Also need to address what to do
with multiple threads in a workspace (probably using thread_id somehow)
and what to do with error messages in terms of marking the line of the
error.