BREAKING CHANGE: Change behaviour of invert_chord
This commit does two things:

  a) Guards against an error calling rings with out of range index
ranges
  b) Makes a breaking change to the behaviour of negative inversions

Prior to this, the implementation of `invert_chord`, when given an
argument of say -1 for the no. of inversions would do this:

```
puts invert_chord([60, 64, 67], -1) # => [64,67,48]
puts invert_chord([60, 64, 67], -2) # => [67,48,52]
```

Musically speaking, it makes more sense to move the top voice down an
octave first, so that the chord is voiced as closely together as
possible i.e. `[:g3, :c4, :e4]` - this voicing is skipped over by the
previous implementation.

See added tests for more details.