Add __run_tests method for running tests from GUI
Having a workspace with the following contents

```
__run_tests
```

and hitting the "Run" button will run the test suite and report the
results back the the GUI.

The formatting is not ideal at the moment. Without TCP a concatenated output
is too long for the OSC protocol we for comms between the Ruby server
and the GUI.

The comments have been left in as a suggestion for an output format. The
guard against `Exception` objects with `nil` backtraces is to facilitate
the option of displaying the test results in the GUIs error window.

NB some tests fail because they are being run directly with Ruby instead
of Rake. This means that files have to be required explicitly which is a
fix up job for a future commit.