Convert tests to Minitest 5 and vendor minitest-5.8.1
Previously we had to consider 3 different testing frameworks:

1) Test::Unit (which worked if you had the test-unit gem installed)
2) Minitest4 (incl. in Ruby <= 1.9.3 - has Test::Unit compatibility mode)
2) Minitest5 (incl. in Ruby >= 2.0.0 - has Test::Unit compatibility mode)

Because our tests were specified with

```
class FooTest < Test::Unit::TestCase
```

and we weren't vendoring a specific testing gem, this meant that the
tests would be run through one of those 3 systems depending on the host
platform and whether you ran through `rake` or `ruby`. This lead to
annoying inconsistencies between Travis and at least my dev environment,
if not others.

Given that we are also aiming to be able to run tests from within Sonic
Pi itself with the `__run_tests` method, it seems sensible to
rationalise on one testing framework. Given that either `test-unit` or
`minitest` require vendoring a version to ensure stability there's not
much in it, but I've opted for `minitest` here as it has zero depencies
and is included in the Ruby stdlib (whereas `test-unit` is more
associated with Rails).

Because we run Ruby 1.9.3 on the Raspberry Pi and 2.1.2 on other
platforms, that means you have to accomodate `minitest v4` _and_
`minitest v5`. They made breaking changes to the API between these
versions with the suggestion of using a backport gem to ensure
compatibility -
[minitest-test](https://github.com/blowmage/minitest-test). The issue
arises when we try to load our vendored gems and any that are installed
locally. It seemed to me that a minitest was appearing elsewhere in the
`$LOAD_PATH` inside our vendor dir and this was causing the backport to
fail. For me, the safest option is to vendor the gem and avoid
ambiguity.

With this commit I now have tests running via `rake` and executing them
directly with `ruby path_to_test.rb` for `ruby` versions `1.9.3` and
`2.1.2`, with and without `minitest` installed locally. The tests also
run and pass when run from the GUI with `__run_tests`