WIP: move to QPlainTextEdit for logging
Apparently `QPlainTextEdit` is more suited to logging performance wise
that the `QTextEdit` we are currently using.

Sources:

http://stackoverflow.com/a/6914301/2618015
http://stackoverflow.com/questions/17466046/qtextedit-vs-qplaintextedit

This commit swaps out the widget but loses the formatting. Apparently it
does work with HTML and stylesheets as it's designed for coloured
logging.

Formatting is probably best achieved with either `setCurrentCharFormat`
or [QSyntaxHighlighter](http://doc.qt.io/qt-4.8/qsyntaxhighlighter.html)