Add styling back in to QPlainTextEdit log widget
Text is styled in a variety of ways - stylesheets, `QTextFormat` and
`QFont` as well as setting properties on the `QPlainTextEdit` itself and
the parent `QWidget`.

There were some methods added into to the log class to maintain some
sort of compatibility between `QTextEdit` and `QPlainTextEdit`.

Also this commit seems to have fixed a bug to do with setting the
initial text colouring inside the `toggleDarkMode` function of
`mainwindow.cpp`. The stylesheet settings here called for the the text
to be coloured grey on grey when in light mode, but for some reason the white foreground
text took precedent when a `QTextEdit` was used. This is fixed now
anyway.