Synthdefs - ensure basic samplers don't kill themselves at the wrong time
The current implementation of the basic sample synths
use a line ugen to self-kill. However, the logic
relied on the rate which had been modified based on
the sample's sample-rate. However, for that particular
logic to work correctly the rate must be the logical 
rate the user specified. This has now been fixed.