Add translations info to README
This puts the fine work of the translators right in the README. This is
partly in response to #737 where we had lost track of @nicoder's ongoing
French translation. Until we can move to a better platform like
transifex or similar, this table is a way of coordinating which
translations are in progress.