Fix tb303 and mono_player :cutoff_decay docs
Previously, both the :cutoff_decay and :cutoff_attack docstrings for these
synths said they referred to the Amount of time (in beats) for sound to reach
full cutoff value. This change corrects this by changing :cutoff_decay to refer
to the amount of time for sound to move from full cutoff value (cutoff attack
level) to the cutoff sustain level.