Document synth design process for contributors
The synth design process requires a knowledge of Overtone. Up to now
this process was also difficult because the folder structure did not
match up to the namespaces used in the synth definition files.

This commit moves them into the necessary folder structure and outlines
the basic process for running an Overtone repl with the Sonic Pi
synthdefs available on the classpath. Hopefully this should make it
(slightly) easier for contributors.