Server - optimise string allocation for OSC messages
In Ruby < 2.3 string allocation causes strain on 
the CPU for both object allocation and subsequent
GC cycles. This can be alleviated in v2.2+ with
calls to .freeze but this version of Ruby isn't
yet available on all binaries. Therefore, achieve
the same result by caching the common OSC path
strings as frozen instance variables and re-using
them across calls to the scsynth OSC API.