Add breakbeat to loop sample list
I was looking at the help documentation just now, specifically the 'sounds for
looping' section of the samples tab. When I started typing in the editor to
reference one of the loop samples, I noticed that the sample :loop_breakbeat
was in the auto-complete list but not in the documentation.
I realised also that this meant that the breakbeat loop sample was not being
shown as part of the output of the 'sample_names' and 'all_sample_names' fns.

This update adds the breakbeat sample to the list of sample metadata, in order
to make it visible in the documentation and the sample name fns.