Fix runtime error due to incorrect validations
Previously, it looks like the 'hard' and 'stereo_width' opts of the piano synth
were being validated against the 'vel' opt instead.
This may have been what was causing a run time error as reported here:
https://groups.google.com/forum/#!msg/sonic-pi/X3IuAdR4slE/XUmZLOjkCgAJ
when trying to use the stereo_width parameter.
This update changes the validations to refer to the correct symbols.