Add po4a based framework for handling tutorial translations as PO files.
The existing translations can be migrated to this new format, but this will
require someone knowing the language to verify the migration result.

To migrate 'de' translation, run 'make LANGUAGES=de import-translations'
and check the content of sonic-pi-tutorial-de.po and
sonic-pi-tutorial-de-failed.po.  The migration require the file structure of
the English original and the translated file to match.  If the migration
have problems, try modifying the translated file structure to match the
English version.