Enable Octaver Effect
This has been languishing since v2.2 but it's ready to go now.

I'd originally thought that the full wave rectifier effect (`abs` in
Overtone) could be used to do transpositions other than an octave but
I'd been mistaking it for another effect. I've gone ahead and
implemented a full range "cheap" transposition in the `whammy` effect so
that leaves this open to fulfil its destiny of fuzzy octave goodness.

The super octave (1 octave up) is produced by `abs` and has a good
Wikipedia article here: https://en.wikipedia.org/wiki/Octave_effect

The sub and subsub are produced by flip flops (`toggle-ff` in Overtone)
and mimic the classic Boss OD-2 guitar pedal.

Both of these mess with the original waveform to some extent so they add
fuzz and artifacts which can be quite pleasant if you're in the right
mood for them.