Make tanh krunchier by default
This commit scales the user input to allow for a krunchier default
setting of tanh. Given the mathematical sounding name, the default
effect was quite subtle which I imagine would lead people to wonder what
it was supposed to be doing. This commit scales the user choice by 2x
resulting in a more noticable default. Aesthetically, this is a breaking
change from v2.9 but I would argue that its not a massive one.

Generated pdf to follow as I'm on a plane and don't have graphviz
installed.