Add guitar synth
This commit adds a simple guitar synth based on the `pluck` UGen, which
in turn uses Karplus-Strong synthesis. This method of synthesis starts
with a noisy input (in this case `pink-noise`) and then filters it
slightly whilst feeding it back on itself with a delay buffer.

It sounds a bit like a guitar, but could also pass for a harpsichord or
a harp. Maybe even a thumb piano for high values of `coef`.

Because this uses a delay buffer internally there is the possibility
that it could become unstable if the memory runs out.

Also no pdf is included for the synthdef as I don't have graphviz
installed at the time of commiting.