WIP: add use_swing/with_swing methods
This allows for a global timing adjustment, similar to the existing
use_tuning but for time!

The reason this is work in progress is that currently it assumes the
length of a musical beat is 1. Given that, it scales the middle of the
beat to be whatever you set for the value of `use_swing`.

For example, with `use_swing 0.75` this will stretch the middle of the
beat to be at 0.75 instead of 0.5

For this to be finished it needs tests and to be able to cope with beat
lengths other than 1