Don't print rest notes if synth logging is off
Recently it was noticed that even if synth logging was turned off via the GUI's preferences, rest notes were still being printed to the log. This was most likely due to changes made in 4323434. A check has now been re-introduced to make sure rest notes are once again only printed to the log if the relevant TL variable is set to false.