WIP: I don't think this will work
Here's the issue with swing. You have long-short long-short (say [0.6,0.4])
according to a beat length. There also exists the possibility
of using other rhythms like 0.25 or 0.33 or 0.66. If these were combined
with one of the longs then it would be necessary to scale all rhythms
according to whether they fell before or after the midpoint of the
swing. This is doable but would sound janky for anything other than the
most basic swing rhythms (1 and 0.5 based rhythms). It would be
misleading to introduce it as a feature when it wasn't intelligent
enough to know what to do with those other rhythms.

Add to this that it wouldn't be able to cope with tempo changes or with
runs that didn't start on zero and I think this isn't a goer.