Add vowel fx
This is a second attempt at adding a vowel fx like Tidal, this time
using the `b-band-pass` ugen. This was based on code from this page

http://sccode.org/1-4Uz

The figures for the vowels come from the FormantTable UGen which is part
of the sc3 plugins. I think these probably came from CSound originally
as that's where everyone seems to copy from for the formants.

Getting `b-band-pass` to match the volume of the unfiltered input
required a large pre and post amp fudge. Also I found it sounded the
same as the SuperCollider version if I summed several filters rather
than using mix as in the SC code.