Synths - introduce new optional method "munge_opts"
If the munge_opts method is overridden in a child
class, then it will be called immediately before
sending any opts to scsynth as part of a synth/fx
trigger. 

The munge_opts fn takes a hash of args and must
also return a hash of args.

Note that this fn is called *after* any GUI logging
so the log won't reflect any munging that happens
here.

FAO @xavriley