Update mac installation guide with required programs
When compiling the ruby files, you'll have to make sure that both cmake and pkg-config are installed on the system. If they are not, the rugged package will not be installed. The error it throws is directly followed by the other warnings the compile-extensions.rb file throws, making it easy to overlook this issue.