sonic-pi-server: add monitor command
Allow a local clients to request to subscribe to OSC messages normally
sent to GUI. This is useful for third-party clients to integrate with
the standard sonic-pi runtime.

The (example) motivation for this change is to allow
Widdershin/sonic-pi-cli to see errors (and other results) from spider
eval. That in turn would allow showing errors and when using vim as an
interface to sonic-pi via dermusikman/sonicpi.vim.