Add new function fx_names
This adds a new function to return a ring of the names of all user facing FX.
The list is constructed by 1) filtering @@synth_infos for FX where user_facing? = true and the name doesn't include 'replace_', and 2) formatting these names to match the FX symbols as used by the user.