GUI - make window show on correct screen in fullscreen
This update gets the screen number of the desktop that most contains the main window widget (eg 'currentScreen') before switching to fullscreen, and then moves the window to appear at the screen indicated by that index in qApp->screens().
This prevents the window from defaulting to one particular screen all the time (eg always appearing on an external monitor instead of a laptop's screen if fullscreen mode was activated from the laptop screen).