GUI - remove logic which supported debian's structure
The existing logic to support debian's re-structuring
of Sonic Pi's folders made Sonic Pi behave *differently*
depending on where it was launched from. This 
meant that on Raspbian all the tests passed, the, 
test binary launched successfully, but when it was
built into a deb and tested it failed to boot due
to a bug in the Raspbian-specific code.

The app should *not* behave differently depending 
on where it was launched from. This produces
surprising and hard-to-debug behaviour.

Instead we should look for another solution - 
perhaps something that requires specific flags
to be set on boot.

@hzulla: For now, this logic is being removed. 
We need to re-implement it to re-introduce support 
for Debian :-)