Switch OSC implementation to fast_osc gem
```
ENCODING TEST
Calculating -------------------------------------
            fast_osc    48.125k i/100ms
             samsosc    19.775k i/100ms
            osc-ruby     8.057k i/100ms
-------------------------------------------------
            fast_osc    847.515k (±20.4%) i/s -      4.091M
             samsosc    264.927k (±18.8%) i/s -      1.285M
            osc-ruby     86.697k (±16.9%) i/s -    427.021k
DECODING TEST
Calculating -------------------------------------
            fast_osc    71.114k i/100ms
             samsosc    19.196k i/100ms
            osc-ruby     3.208k i/100ms
-------------------------------------------------
            fast_osc      1.928M (±17.8%) i/s -      9.387M
             samsosc    247.874k (±13.1%) i/s -      1.229M
            osc-ruby     31.996k (±14.7%) i/s -    157.192k
```

After several months of research, this commit includes a new OSC
implementation for Ruby using a C extension for speed. This includes the
`rtosc` library in the source and does not require any other
dependencies.

Open Sound Control (OSC) encoding and decoding is one of the major
hotspots for Sonic Pi and despite some intense and largely succesful
work in optimising the pure Ruby implementation, I felt that this could
be improved upon.

The performance matters not only on the Raspberry Pi (in particular the
model B+) but also for future applications that may use more OSC
messaging in general. At the moment we only use OSC messages to/from the
GUI and to/from `scsynth`. This higher performance implementation paves
the way for some other "real time" applications of OSC in future.

As a separate gem, it is available to the wider Ruby community for use
in other OSC applications where performance is critical.