Sound - short-circuit on should_trigger? at last opportunity
Wait until the last possible opportunity before 
checking args_h whether to trigger. This ensures
that all arg munging has completed. This is
essential for a maintaining a consistent number of 
stateful changes such as consumptions of rands
(which is the main point of using the on: opt)

For example, the following now both consume 1 rand:

sample :foo, onset: :rand, on: true
sample :foo, onset: :rand, on: false