Core - ensure user thread locals have correct defaults on reset
The fn reset nukes the user’s thread locals. Unfortunately this also had the side effect of nuking the TLs which were assumed to have known default values (as set when the Run is created).

Now, we always set the TLs which should *always* have a default on reset using the same method as the Run creation code.